refactor: streamline user management queries by removing redundant code and enhancing clarity
