+++ src/components/dashboard/sidebar.tsx
@@ -3,7 +3,7 @@
 import { logoutAction } from '@/app/actions/auth';
 import { Button } from '@/components/ui/button';
 import { ScrollArea } from '@/components/ui/scroll-area';
-import { useAuth } from '@/lib/api/queries/useAuth';
+import { useAuth } from '@/lib/api/queries/use-auth';
 
 import { cn } from '@/lib/utils';
 import {
